requirements.txt
# requirements.txt
Flask==2.1.1
Werkzeug==2.0.3
psycopg2-binary==2.9.9
